requirements.txt
streamlit==1.40.0
torch==2.2.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
python-dotenv==1.0.1
openai==1.10.0
ollama==0.1.6
pillow==10.2.0
jupyter==1.0.0
ffmpeg-python==0.2.0
mlflow==2.16.2